requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LLM & Embeddings
openai==1.9.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# Vector Database
chromadb==0.4.18

# LangGraph Workflow
langgraph==0.0.26
langchain==0.1.14
langchain-core==0.1.37

# Additional Dependencies
numpy==1.24.3
requests==2.31.0
google-generativeai==0.3.2
